add shortcut key in WindowButtons.vue
